CLIs/explainclip/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "explainclip"
version = "0.1.0"
description = "Production-quality animated explainer videos from the command line"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MotusAI", email = "hello@motusai.com" },
]
keywords = ["manim", "animation", "explainer", "video", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "manim>=0.18.0",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff>=0.4",
    "mypy",
]

[project.scripts]
explainclip = "explainclip.cli:main"

[project.urls]
Homepage = "https://github.com/motusai/explainclip"
Documentation = "https://github.com/motusai/explainclip#readme"
Repository = "https://github.com/motusai/explainclip"
Issues = "https://github.com/motusai/explainclip/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.hatch.build.targets.wheel]
packages = ["src/explainclip"]
